Add default seqview
