+++ test/fastq/record.jl
@@ -1,0 +1,3 @@
+# Only using empty records here
+@testset "Basic properties" begin
+end
